Fixes issue with gcc and std::sqrtf

+++ Source/Engine/StratusMath.h
@@ -417,7 +417,7 @@ namespace stratus {
         float dy = std::max<float>(aabb.vmin.v[1] - point.y, std::max<float>(0.0f, point.y - aabb.vmax.v[1]));
         float dz = std::max<float>(aabb.vmin.v[2] - point.z, std::max<float>(0.0f, point.z - aabb.vmax.v[2]));
 
-        return std::sqrtf(dx * dx + dy * dy + dz + dz);
+        return std::sqrt(dx * dx + dy * dy + dz + dz);
     }
 
     // These are the first 512 values of the Halton sequence. For more information see:
